src/requirements.txt
# ====================== 核心依赖 ======================
# Web框架
tornado==6.4

# 数据库
mysql-connector-python==8.3.0
sqlalchemy==2.0.25
alembic==1.13.1

# 缓存
redis==5.0.1

# 消息队列（EDA事件总线）
pika==1.3.2

# 配置管理
python-dotenv==1.0.0

# 数据验证
pydantic==1.10.13

# 日志
loguru==0.7.2

# 安全
pyjwt==2.8.0
bcrypt==4.1.2

# ====================== 工具依赖 ======================
# 时间处理
python-dateutil==2.8.2
pytz==2024.1

# ====================== 开发依赖 ======================
# 测试
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# 代码质量
black==24.1.1
flake8==7.0.0
isort==5.13.2